requirements.txt
azure-ai-vision-imageanalysis
fastapi
openai
python-dotenv
uvicorn